requirements.txt
google-adk[all]
fastapi
uvicorn[standard]
streamlit
python-dotenv
pydantic
